os2ws2 test day1

+++ include/dataflow.h
@@ -15,7 +15,7 @@ public:
 	isl_union_map *GetTimeMap();
 	isl_union_map *GetSpaceTimeMap();
 	isl_union_set *GetDomain();
-	int GetDomainSize();
+	double GetDomainSize();
 	isl_union_map *GetAccess(std::string tensor_name,
 		AccessType type);
 	isl_union_set *GetSpaceDomain();
@@ -27,31 +27,31 @@ public:
 		bool space_is_range=true, unsigned time_distance = 1, 
 		bool time_is_range = true, bool include_self = false);
 	isl_union_map * MapSpaceTimeToAccess(std::string tensor_name, AccessType type);
-	int GetPENum();
-	int GetTotalTime();
+	double GetPENum();
+	double GetTotalTime();
 	// following are functions that perform dataflow analysis
-	int GetUniqueVolume(std::string tensor_name, AccessType type,
+	double GetUniqueVolume(std::string tensor_name, AccessType type,
 		isl_union_map* space_time_to_neighbor);
-	int GetTotalVolume(std::string tensor_name, AccessType type);
+	double GetTotalVolume(std::string tensor_name, AccessType type);
 	double GetReuseFactor(std::string tensor_name, AccessType type,
 		isl_union_map* space_time_to_neighbor);
 	double GetTemporalReuseVolume(std::string tensor_name, AccessType type);
 	double GetSpatialReuseVolume(std::string tensor_name, AccessType type, isl_union_map* stt_neighbor);
-	int GetMacNum(int mac_per_instance = 1);
-	int GetMacNumPerPE(int mac_per_instance = 1);
-	int GetActivePENum();
+	double GetMacNum(int mac_per_instance = 1);
+	double GetMacNumPerPE(int mac_per_instance = 1);
+	double GetActivePENum();
 	double GetAverageActivePENum();
-	int GetIngressDelay(isl_union_map* space_time_to_neighbor , std::string tensor_name = "");
-	int GetEgressDelay(isl_union_map* space_time_to_neighbor, std::string tensor_name = "");
-	int GetComputationDelay();
-	int GetDelay(isl_union_map* space_time_to_neighbor);
-	int GetL1Read(std::string tensor_name, AccessType type);
-	int GetL1Write(std::string tensor_name, AccessType type);
-	int GetL2Read(std::string tensor_name, AccessType type,
+	double GetIngressDelay(isl_union_map* space_time_to_neighbor , std::string tensor_name = "");
+	double GetEgressDelay(isl_union_map* space_time_to_neighbor, std::string tensor_name = "");
+	double GetComputationDelay();
+	double GetDelay(isl_union_map* space_time_to_neighbor);
+	double GetL1Read(std::string tensor_name, AccessType type);
+	double GetL1Write(std::string tensor_name, AccessType type);
+	double GetL2Read(std::string tensor_name, AccessType type,
 		isl_union_map* space_time_to_neighbor);
-	int GetL2Write(std::string tensor_name, AccessType type,
+	double GetL2Write(std::string tensor_name, AccessType type,
 		isl_union_map* space_time_to_neighbor);
-	int GetEnergy(isl_union_map* space_time_to_neighbor);
+	double GetEnergy(isl_union_map* space_time_to_neighbor);
 
 	Dataflow copy() const;
 
@@ -60,7 +60,7 @@ private:
 	PEArray _pe;
 	Mapping _mp;
 
-	int convert_upwqp_to_int(isl_union_pw_qpolynomial *upwqp);
+	double convert_upwqp_to_int(isl_union_pw_qpolynomial *upwqp);
 };
 
 } // namespace TENET
